+++ Source/QualityBionicsContinued/Constants.cs
@@ -1,6 +1,0 @@
-namespace QualityBionicsContinued;
-
-internal static class Constants
-{
-    internal const string Id = "QualityBionicsContinued";
-}