+++ Source/QualityBionicsRemastered/Constants.cs
@@ -1,0 +1,6 @@
+namespace QualityBionicsRemastered;
+
+internal static class Constants
+{
+    internal const string Id = "QualityBionicsRemastered";
+}