+++ Source/QualityBionicsRemastered/Patch/BackCompatibility_GetBackCompatibleType.cs
@@ -1,7 +1,7 @@
 
 using System;
 using HarmonyLib;
-using QualityBionicsContinued;
+using QualityBionicsRemastered;
 using Verse;
 
 [HarmonyPatch(typeof(BackCompatibility), nameof(BackCompatibility.GetBackCompatibleType))]
@@ -10,7 +10,10 @@ internal static class BackCompatibility_GetBackCompatibleType
     private static void Postfix(ref Type __result, string providedClassName)
     {
         // This lets us load our old settings without RimWorld producing an error
-        if (providedClassName == "QualityBionicsContinued.QualityBionicsSettings" || providedClassName == "QualityBionics.QualityBionicsSettings")
+        if (providedClassName == "QualityBionicsContinued.QualityBionicsSettings" || 
+            providedClassName == "QualityBionics.QualityBionicsSettings" ||
+            providedClassName == "QualityBionicsContinued.Settings" ||
+            providedClassName == "QualityBionicsRemastered.Settings")
             __result = typeof(Settings);
     }
 }